add: common components for chat
